requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==4.4.2
Django==5.0.7
django-environ==0.11.2
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy==2.0.0
pillow==10.4.0
proglog==0.1.10
psycopg2-binary==2.9.9
requests==2.32.3
sqlparse==0.5.0
tqdm==4.66.4
urllib3==2.2.2
